--- cmd/Flag.go
@@ -1,16 +1,0 @@
-package main
-
-type Flag struct {
-	set   bool
-	value string
-}
-
-func (sf *Flag) Set(x string) error {
-	sf.value = x
-	sf.set = true
-	return nil
-}
-
-func (sf *Flag) String() string {
-	return sf.value
-}
